+++ cycletls/tests/integration/disable_redirect_test.go
@@ -14,10 +14,10 @@ import (
 func TestRedirectEnabled(t *testing.T) {
 
 	client := cycletls.Init()
-	resp, err := client.Do("https://google.com", cycletls.Options{
+	resp, err := client.Do("https://ssl.com", cycletls.Options{
 		Body:      "",
-		Ja3:       "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0",
-		UserAgent: "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:87.0) Gecko/20100101 Firefox/87.0",
+		Ja3:       "771,4865-4866-4867-49195-49199-49196-49200-52393-52392-49171-49172-156-157-47-53,0-23-65281-10-11-35-16-5-13-18-51-45-43-27-21,29-23-24,0",
+		UserAgent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/83.0.4103.106 Safari/537.36",
 	}, "GET")
 	if err != nil {
 		log.Print("Request Failed: " + err.Error())
@@ -31,10 +31,10 @@ func TestRedirectEnabled(t *testing.T) {
 func TestRedirectDisabled(t *testing.T) {
 
 	client := cycletls.Init()
-	resp, err := client.Do("https://google.com", cycletls.Options{
+	resp, err := client.Do("https://ssl.com", cycletls.Options{
 		Body:            "",
-		Ja3:             "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0",
-		UserAgent:       "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:87.0) Gecko/20100101 Firefox/87.0",
+		Ja3:             "771,4865-4866-4867-49195-49199-49196-49200-52393-52392-49171-49172-156-157-47-53,0-23-65281-10-11-35-16-5-13-18-51-45-43-27-21,29-23-24,0",
+		UserAgent:       "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/83.0.4103.106 Safari/537.36",
 		DisableRedirect: true,
 	}, "GET")
 	if err != nil {