+++ cycletls/tests/integration/header_order_test.go
@@ -5,17 +5,17 @@ package cycletls_test
 
 import (
 	//"fmt"
+	"encoding/json"
 	"github.com/PuerkitoBio/goquery"
 	"log"
 	"strings"
 	"testing"
-	"encoding/json"
 
 	cycletls "github.com/Danny-Dasilva/CycleTLS/cycletls"
 )
 
 type HttpBinHeaders struct {
-    Headers   map[string]string      
+	Headers map[string]string
 }
 
 //TODO rewrite this so its not reliant on goquery
@@ -35,7 +35,7 @@ func TestDefaultHeaderOrder(t *testing.T) {
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(string(resp.Body)))
 
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal("Document read error",err)
 	}
 
 	headername := doc.Find(".headername").Text()
@@ -62,7 +62,7 @@ func TestCustomHeaderOrder(t *testing.T) {
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(string(resp.Body)))
 
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal("Document read error",err)
 	}
 
 	headername := doc.Find(".headername").Text()
@@ -88,7 +88,7 @@ func TestCustomHeaderOrderFailure(t *testing.T) {
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(string(resp.Body)))
 
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal("Document read error",err)
 	}
 
 	headername := doc.Find(".headername").Text()
@@ -99,14 +99,13 @@ func TestCustomHeaderOrderFailure(t *testing.T) {
 	}
 }
 
-
 func TestCustomHeadersDefaultOrder(t *testing.T) {
 	client := cycletls.Init()
 	resp, err := client.Do("https://pgl.yoyo.org/http/browser-headers.php", cycletls.Options{
-		Body:        "",
-		Ja3:         "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0",
-		UserAgent:   "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:87.0) Gecko/20100101 Firefox/87.0",
-		Headers:     map[string]string{"test1": "value1", "test2": "value2"},
+		Body:      "",
+		Ja3:       "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0",
+		UserAgent: "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:87.0) Gecko/20100101 Firefox/87.0",
+		Headers:   map[string]string{"test1": "value1", "test2": "value2"},
 	}, "GET")
 	if err != nil {
 		log.Print("Request Failed: " + err.Error())
@@ -114,7 +113,7 @@ func TestCustomHeadersDefaultOrder(t *testing.T) {
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(string(resp.Body)))
 
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal("Document read error",err)
 	}
 
 	headername := doc.Find(".headername").Text()
@@ -140,7 +139,7 @@ func TestCustomHeadersCustomOrder(t *testing.T) {
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(string(resp.Body)))
 
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal("Document read error",err)
 	}
 
 	headername := doc.Find(".headername").Text()
@@ -153,10 +152,11 @@ func TestCustomHeadersCustomOrder(t *testing.T) {
 func TestCustomHeaders(t *testing.T) {
 	client := cycletls.Init()
 	resp, err := client.Do("https://httpbin.org/headers", cycletls.Options{
-		Body:        "",
-		Ja3:         "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0",
-		UserAgent:   "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:87.0) Gecko/20100101 Firefox/87.0",
-		Headers:     map[string]string{"foo": "bar"},
+		Body:      "",
+		Ja3:       "771,4865-4867-4866-49195-49199-52393-52392-49196-49200-49162-49161-49171-49172-51-57-47-53-10,0-23-65281-10-11-35-16-5-51-43-13-45-28-21,29-23-24-25-256-257,0",
+		UserAgent: "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:87.0) Gecko/20100101 Firefox/87.0",
+		Headers:   map[string]string{"foo": "bar"},
+		Timeout: 30,
 	}, "GET")
 	if err != nil {
 		log.Print("Request Failed: " + err.Error())
@@ -164,7 +164,7 @@ func TestCustomHeaders(t *testing.T) {
 	var result HttpBinHeaders
 	err = json.Unmarshal([]byte(resp.Body), &result)
 	if err != nil {
-		log.Fatal(err)
+		log.Fatal("Unmarshal error",err)
 	}
 	if result.Headers["Foo"] != "bar" {
 		t.Fatalf("Headers not applied")